// crud/src/app/models/cliente.model.ts
export interface Cliente{
    codigoCliente: number;
    codigoCidade: number;
    nomeCliente: string;
    enderecoCliente: string;
    bairroCliente: string;
    emailCliente: string;
    descricaoCidade: string;
    numeroCliente: string;
    cepCliente: string;
    cpfCnpjCliente: string;
    telefoneCliente: string;
}